Rework IdeResponse into monad with deferred type
