+++ hie-plugin-api/Haskell/Ide/Engine/MonadTypes.hs
@@ -10,14 +10,12 @@
 module Haskell.Ide.Engine.MonadTypes
   (
   -- * All the good types
-    module Haskell.Ide.Engine.PluginTypes
-  , module Haskell.Ide.Engine.MultiThreadState
+    module Haskell.Ide.Engine.MultiThreadState
   , module Haskell.Ide.Engine.PluginsIdeMonads
   , module Haskell.Ide.Engine.GhcModuleCache
   , module Haskell.Ide.Engine.ModuleCache
   ) where
 
-import           Haskell.Ide.Engine.PluginTypes
 import           Haskell.Ide.Engine.MultiThreadState
 import           Haskell.Ide.Engine.PluginsIdeMonads
 import           Haskell.Ide.Engine.GhcModuleCache
