+++ hie-plugin-api/hie-plugin-api.cabal
@@ -28,7 +28,6 @@ library
                        Haskell.Ide.Engine.MultiThreadState
                        Haskell.Ide.Engine.PluginDescriptor
                        Haskell.Ide.Engine.PluginsIdeMonads
-                       Haskell.Ide.Engine.PluginTypes
                        Haskell.Ide.Engine.PluginUtils
   build-depends:       base >= 4.9 && < 5
                      , Diff
